Removed old code
