--- appmanager2/default-app/smplatform-welcome.html
@@ -1,18 +1,0 @@
-<!doctype html>
-
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-
-    <title>Welcome to SMPlatform</title>
-    <meta name="description" content="Welcome to SMPlatform">
-    <meta name="author" content="ItalyPaleAle">
-</head>
-
-<body>
-    <h1>Welcome to SMPlatform</h1>
-    <p>One day this needs to be filled out</p>
-</body>
-
-</html>
