+++ webserver/nginx-template/site.conf
@@ -8,15 +8,15 @@ server {
 
     # Configure logging
     access_log off;
-    error_log {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/nginx-error.log error;
+    error_log {{.AppRoot}}sites/{{.Item.Domain}}/nginx-error.log error;
 
     # TLS
-    ssl_certificate {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/tls/certificate.pem;
-    ssl_certificate_key {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/tls/key.pem;
-    ssl_dhparam {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/tls/dhparams.pem;
+    ssl_certificate {{.AppRoot}}sites/{{.Item.Domain}}/tls/certificate.pem;
+    ssl_certificate_key {{.AppRoot}}sites/{{.Item.Domain}}/tls/key.pem;
+    ssl_dhparam {{.AppRoot}}sites/{{.Item.Domain}}/tls/dhparams.pem;
 
     # Webroot
-    root {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/www;
+    root {{.AppRoot}}sites/{{.Item.Domain}}/www;
     index index.html index.htm;
 
 {{if .Item.ClientCaching}}
@@ -49,7 +49,7 @@ server {
 
     # Configure logging
     access_log off;
-    error_log {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/nginx-error.log error;
+    error_log {{.AppRoot}}sites/{{.Item.Domain}}/nginx-error.log error;
 
     # Redirect to the HTTPS website
     return 301 https://{{.Item.Domain}}$request_uri;
@@ -66,12 +66,12 @@ server {
 
     # Configure logging
     access_log off;
-    error_log {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/nginx-error.log error;
+    error_log {{.AppRoot}}sites/{{.Item.Domain}}/nginx-error.log error;
 
     # TLS
-    ssl_certificate {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/tls/certificate.pem;
-    ssl_certificate_key {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/tls/key.pem;
-    ssl_dhparam {{.AppRoot}}sites/{{uuidToString .Item.SiteID}}/tls/dhparams.pem;
+    ssl_certificate {{.AppRoot}}sites/{{.Item.Domain}}/tls/certificate.pem;
+    ssl_certificate_key {{.AppRoot}}sites/{{.Item.Domain}}/tls/key.pem;
+    ssl_dhparam {{.AppRoot}}sites/{{.Item.Domain}}/tls/dhparams.pem;
 
     # Redirect to the canonical host
     return 301 https://{{.Item.Domain}}$request_uri;
