+++ webserver/nginx.go
@@ -17,6 +17,9 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 package webserver
 
 import (
+	"bytes"
+	"errors"
+	"io/ioutil"
 	"log"
 	"os"
 	"os/exec"
@@ -24,35 +27,28 @@ import (
 	"text/template"
 
 	"smplatform/appconfig"
-	"smplatform/db"
+	"smplatform/state"
 	"smplatform/utils"
 
 	"github.com/gobuffalo/packr/v2"
-	"github.com/gofrs/uuid"
 )
 
-/* Constants */
-
+// List of template files
 var templateFiles = [4]string{"nginx.conf", "mime.types", "site.conf", "default-site.conf"}
 
-/* NginxConfig struct */
+// ConfigData is a map of each configuration file and its content
+type ConfigData map[string][]byte
 
 // NginxConfig creates the configuration for nginx
 type NginxConfig struct {
-	appRoot       string
-	nginxConfPath string
-	logger        *log.Logger
-	templates     map[string]*template.Template
+	logger    *log.Logger
+	templates map[string]*template.Template
 }
 
 // Init initializes the object and loads the templates from file
 func (n *NginxConfig) Init() error {
 	// Logger
 	n.logger = log.New(os.Stdout, "nginx: ", log.Ldate|log.Ltime|log.LUTC)
-
-	// Init properties from env vars
-	n.appRoot = appconfig.Config.GetString("appRoot")
-	n.nginxConfPath = appconfig.Config.GetString("nginx.configPath")
 
 	// Load the templates
 	n.templates = make(map[string]*template.Template, len(templateFiles))
@@ -63,75 +59,197 @@ func (n *NginxConfig) Init() error {
 	return nil
 }
 
-// ResetConfiguration resets the nginx configuration so there's only the default website
-func (n *NginxConfig) ResetConfiguration() error {
-	// Ensure the folder exists
-	if err := utils.EnsureFolder(n.nginxConfPath); err != nil {
-		return err
-	}
-
-	// Clear the contents of the folder
-	if err := utils.RemoveContents(n.nginxConfPath); err != nil {
-		return err
-	}
+// DesiredConfiguration builds the list of files for the desired configuration for nginx
+func (n *NginxConfig) DesiredConfiguration(sites []state.SiteState) (config ConfigData, err error) {
+	config = make(ConfigData)
 
 	// Basic webserver configuration
-	if err := n.writeConfigurationFile("nginx.conf", "nginx.conf", nil); err != nil {
-		return err
+	config["nginx.conf"], err = n.createConfigurationFile("nginx.conf", nil)
+	if err != nil {
+		return
+	}
+	if config["nginx.conf"] == nil {
+		err = errors.New("Invalid configuration generated for file nginx.conf")
+		return
 	}
 
-	if err := n.writeConfigurationFile("mime.types", "mime.types", nil); err != nil {
-		return err
+	config["mime.types"], err = n.createConfigurationFile("mime.types", nil)
+	if err != nil {
+		return
 	}
-
-	// Create the conf.d folder for websites
-	if err := utils.EnsureFolder(n.nginxConfPath + "conf.d"); err != nil {
-		return err
+	if config["mime.types"] == nil {
+		err = errors.New("Invalid configuration generated for file mime.types")
+		return
 	}
 
 	// Default website configuration
-	if err := n.writeConfigurationFile("conf.d/_default.conf", "default-site.conf", nil); err != nil {
-		return err
+	config["conf.d/_default.conf"], err = n.createConfigurationFile("default-site.conf", nil)
+	if err != nil {
+		return
+	}
+	if config["conf.d/_default.conf"] == nil {
+		err = errors.New("Invalid configuration generated for file conf.d/_default.conf")
+		return
 	}
 
-	return nil
+	// Configuration for each site
+	for _, s := range sites {
+		key := "conf.d/" + s.Domain + ".conf"
+		var val []byte
+		val, err = n.createConfigurationFile("site.conf", s)
+		if err != nil {
+			return
+		}
+		if val == nil {
+			err = errors.New("Invalid configuration generated for file " + key)
+			return
+		}
+		config[key] = val
+	}
+
+	return
 }
 
-// SyncConfiguration ensures that the configuration for the webserver matches the desired state
-func (n *NginxConfig) SyncConfiguration(sites []db.Site) error {
-	// First, reset the configuration
-	err := n.ResetConfiguration()
-	if err != nil {
-		return err
-	}
+// ExistingConfiguration reads the list of files currently on disk, and deletes some extraneous ones already
+func (n *NginxConfig) ExistingConfiguration(sites []state.SiteState) (ConfigData, bool, error) {
+	nginxConfPath := appconfig.Config.GetString("nginx.configPath")
+	existing := make(ConfigData)
+	updated := false
 
-	// Create the configuration file for each website
-	for _, site := range sites {
-		err = n.ConfigureSite(&site)
-		if err != nil {
-			return err
+	// Start with the nginxConfigPath directory
+	files, err := ioutil.ReadDir(nginxConfPath)
+	if err != nil {
+		return nil, false, err
+	}
+	for _, f := range files {
+		name := f.Name()
+		// There should be only 1 directory: conf.d
+		if f.IsDir() {
+			if name != "conf.d" {
+				// Delete the folder
+				updated = true
+				n.logger.Println("Removing extraneous folder", nginxConfPath+name)
+				if err := os.RemoveAll(nginxConfPath + name); err != nil {
+					return nil, false, err
+				}
+			}
+		} else {
+			// There should only be two files: nginx.conf and mime.type
+			if name == "nginx.conf" || name == "mime.types" {
+				var err error
+				existing[name], err = ioutil.ReadFile(nginxConfPath + name)
+				if err != nil {
+					return nil, false, err
+				}
+			} else {
+				// Delete the extraneous file
+				updated = true
+				n.logger.Println("Removing extraneous file", nginxConfPath+name)
+				if err := os.Remove(nginxConfPath + name); err != nil {
+					return nil, false, err
+				}
+			}
 		}
 	}
 
-	return nil
-}
-
-// ConfigureSite creates the configuration for a website
-func (n *NginxConfig) ConfigureSite(site *db.Site) error {
-	if err := n.writeConfigurationFile("conf.d/"+site.SiteID.String()+".conf", "site.conf", site); err != nil {
-		return err
+	// List of files we expect in the conf.d directory
+	existing["conf.d/_default.conf"] = nil
+	for k := range sites {
+		existing["conf.d/"+sites[k].Domain+".conf"] = nil
 	}
 
-	return nil
-}
-
-// RemoveSite removes the configuration for a website
-func (n *NginxConfig) RemoveSite(siteName string) error {
-	if err := os.Remove(n.nginxConfPath + "conf.d/" + siteName + ".conf"); err != nil {
-		return err
+	// Scan the conf.d directory
+	files, err = ioutil.ReadDir(nginxConfPath + "conf.d")
+	if err != nil {
+		return nil, false, err
+	}
+	for _, f := range files {
+		// There shouldn't be any directory
+		if f.IsDir() {
+			// Delete the folder
+			updated = true
+			name := f.Name()
+			n.logger.Println("Removing extraneous folder", nginxConfPath+"conf.d/"+name)
+			if err := os.RemoveAll(nginxConfPath + "conf.d/" + name); err != nil {
+				return nil, false, err
+			}
+		} else {
+			// Expect certain files only
+			key := "conf.d/" + f.Name()
+			_, ok := existing[key]
+			if ok {
+				// We are expecting this file: read it
+				var err error
+				existing[key], err = ioutil.ReadFile(nginxConfPath + key)
+				if err != nil {
+					return nil, false, err
+				}
+			} else {
+				// Delete the extraneous file
+				updated = true
+				n.logger.Println("Removing extraneous file", nginxConfPath+key)
+				if err := os.Remove(nginxConfPath + key); err != nil {
+					return nil, false, err
+				}
+			}
+		}
 	}
 
-	return nil
+	return existing, updated, nil
+}
+
+// SyncConfiguration ensures that the configuration for the webserver matches the desired state
+func (n *NginxConfig) SyncConfiguration(sites []state.SiteState) (bool, error) {
+	nginxConfPath := appconfig.Config.GetString("nginx.configPath")
+	updated := false
+
+	// Generate the desired configuration
+	desired, err := n.DesiredConfiguration(sites)
+	if err != nil {
+		return false, err
+	}
+
+	// Ensure that the required folders exist
+	if err := utils.EnsureFolder(nginxConfPath); err != nil {
+		return false, err
+	}
+	if err := utils.EnsureFolder(nginxConfPath + "conf.d"); err != nil {
+		return false, err
+	}
+
+	// Scan the existing content
+	existing, u, err := n.ExistingConfiguration(sites)
+	updated = updated || u
+	if err != nil {
+		return false, err
+	}
+
+	// Iterate through the desired state looking for missing keys and different files
+	// We're guaranteed that the existing state does not contain any extraenous file already
+	for key, val := range desired {
+		// Check if the file exists already
+		existingVal, ok := existing[key]
+		if ok && existingVal != nil {
+			// Compare the files
+			if bytes.Compare(val, existingVal) != 0 {
+				// Files are different
+				updated = true
+				n.logger.Println("Replacing configuration file", nginxConfPath+key)
+				if err := writeConfigFile(nginxConfPath+key, val); err != nil {
+					return false, err
+				}
+			}
+		} else {
+			// File is missing
+			updated = true
+			n.logger.Println("Creating configuration file", nginxConfPath+key)
+			if err := writeConfigFile(nginxConfPath+key, val); err != nil {
+				return false, err
+			}
+		}
+	}
+
+	return updated, nil
 }
 
 // RestartServer restarts the Nginx server
@@ -149,18 +267,13 @@ func (n *NginxConfig) RestartServer() error {
 // Read all templates
 func (n *NginxConfig) loadTemplates() error {
 	// Packr
-	box := packr.New("Nginx templates", "nginx-template")
+	box := packr.New("Nginx templates old", "nginx-template")
 
 	// Functions to add to the template
 	funcMap := template.FuncMap{
 		// Joins all values of a slice with a string separator
 		"joinList": func(slice []string, separator string) string {
 			return strings.Join(slice, separator)
-		},
-
-		// Converts UUIDs to strings
-		"uuidToString": func(u uuid.UUID) string {
-			return u.String()
 		},
 	}
 
@@ -180,8 +293,8 @@ func (n *NginxConfig) loadTemplates() error {
 	return nil
 }
 
-// Write a configuration to disk
-func (n *NginxConfig) writeConfigurationFile(path string, templateName string, itemData interface{}) error {
+// Create a configuration file
+func (n *NginxConfig) createConfigurationFile(templateName string, itemData interface{}) ([]byte, error) {
 	protocol := "http"
 	if appconfig.Config.GetBool("tls.enabled") {
 		protocol = "https"
@@ -201,7 +314,7 @@ func (n *NginxConfig) writeConfigurationFile(path string, templateName string, i
 		}
 	}{
 		Item:     itemData,
-		AppRoot:  n.appRoot,
+		AppRoot:  appconfig.Config.GetString("appRoot"),
 		Port:     appconfig.Config.GetString("port"),
 		Protocol: protocol,
 		TLS: struct {
@@ -220,15 +333,29 @@ func (n *NginxConfig) writeConfigurationFile(path string, templateName string, i
 	// Get the template
 	tpl := n.templates[templateName]
 
-	// Create the file and execute the template
-	f, err := os.Create(n.nginxConfPath + path)
+	// Execute the template
+	var buf bytes.Buffer
+	if err := tpl.Execute(&buf, tplData); err != nil {
+		return nil, err
+	}
+
+	return buf.Bytes(), nil
+}
+
+// Writes data to a configuration file
+func writeConfigFile(path string, val []byte) error {
+	// Running f.Close() manually to avoid having too many open file descriptors
+	f, err := os.Create(path)
+	defer f.Close()
 	if err != nil {
 		return err
 	}
-	if err := tpl.Execute(f, tplData); err != nil {
+	if err := f.Chmod(0644); err != nil {
 		return err
 	}
-	f.Close()
-
+	_, err = f.Write(val)
+	if err != nil {
+		return err
+	}
 	return nil
 }
