--- webserver2/nginx-template/default-site.conf
@@ -1,56 +1,0 @@
-# Default website
-server {
-    listen 80 default_server;
-
-    # Listen on the domain
-    server_name monitoring;
-
-    # Configure logging
-    access_log off;
-    error_log {{.AppRoot}}sites/_default/nginx-error.log error;
-
-    # Webroot
-    root {{.AppRoot}}sites/_default/www;
-    error_page 404 403 =404 /smplatform-welcome.html;
-
-    # Status pages are proxied to the upstream controller
-    location ~ ^/(status|info) {
-        proxy_set_header X-Forwarded-Proto $scheme;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-        proxy_ssl_verify off;
-        proxy_pass {{.Protocol}}://localhost:{{.Port}};
-    }
-}
-
-{{if .TLS.Enabled}}
-# Default website with TLS
-server {
-    listen 443 ssl http2 default_server;
-
-    # Listen on the domain
-    server_name monitoring;
-
-    # Configure logging
-    access_log off;
-    error_log {{.AppRoot}}sites/_default/nginx-error.log error;
-
-    # Webroot
-    root {{.AppRoot}}sites/_default/www;
-    error_page 404 403 =404 /smplatform-welcome.html;
-
-    # TLS
-    ssl_certificate {{.TLS.Certificate}};
-    ssl_certificate_key {{.TLS.Key}};
-    ssl_dhparam {{.TLS.Dhparams}};
-
-    # Status pages are proxied to the upstream controller
-    location ~ ^/(status|info) {
-        proxy_set_header X-Forwarded-Proto $scheme;
-        proxy_set_header X-Real-IP $remote_addr;
-        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-        proxy_ssl_verify off;
-        proxy_pass {{.Protocol}}://localhost:{{.Port}};
-    }
-}
-{{end}}
